backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
PyJWT
psycopg2-binary
Werkzeug
gunicorn